requirements.txt
Pillow~=10.2.0
Flask~=3.0.2
tensorflow>=2.0.0
numpy~=1.26.4
keras~=3.1.1
requests~=2.31.0
